SVD_recommender/requirements.txt
pandas==2.0.0
numpy==1.24.3
pyodbc==4.0.39
scikit-surprise==1.1.3
Flask==2.3.2
redis==4.5.5